Made a fantastic start at implementing the
STV prototype. At the moment, i have it reading in from a file

correctly, but need to fully sort my input data so that it is a ranked

list and takes into account preferences.

Close to getting the 1st round finished as at the moment the program

returns the candidates name, their first preference votes and their

second preference votes as well as returning the correct quota required

to win a seat in the first round, that can then be recalculated once i

have started my calculation function and redistribution of surplus votes

method.